# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthResponse {
  token: string;
  username: string;
}

export type ToolStatus = (typeof ToolStatus)[keyof typeof ToolStatus];

export const ToolStatus = {
  available: "available",
  issued: "issued",
  missing: "missing",
} as const;

export interface Tool {
  id: number;
  toolId: string;
  name: string;
  category: string;
  status: ToolStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type ToolInputStatus =
  (typeof ToolInputStatus)[keyof typeof ToolInputStatus];

export const ToolInputStatus = {
  available: "available",
  issued: "issued",
  missing: "missing",
} as const;

export interface ToolInput {
  /** @minLength 1 */
  toolId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  status?: ToolInputStatus;
}

export type ToolUpdateStatus =
  (typeof ToolUpdateStatus)[keyof typeof ToolUpdateStatus];

export const ToolUpdateStatus = {
  available: "available",
  issued: "issued",
  missing: "missing",
} as const;

export interface ToolUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  status?: ToolUpdateStatus;
}

export interface IssueInput {
  toolId: string;
  userId: string;
}

export interface ReturnInput {
  toolId: string;
}

export interface ScanInput {
  scannedToolIds: string[];
}

export interface ScanSummary {
  totalScanned: number;
  totalInDb: number;
  correctCount: number;
  missingCount: number;
  extraCount: number;
}

export interface ScanResult {
  correct: Tool[];
  missing: Tool[];
  extra: string[];
  summary: ScanSummary;
}

export interface Stats {
  totalTools: number;
  availableTools: number;
  issuedTools: number;
  missingTools: number;
  recentTransactions: number;
}

export type TransactionAction =
  (typeof TransactionAction)[keyof typeof TransactionAction];

export const TransactionAction = {
  issue: "issue",
  return: "return",
  scan: "scan",
} as const;

export interface Transaction {
  id: number;
  toolId: string;
  /** @nullable */
  toolName?: string | null;
  /** @nullable */
  userId?: string | null;
  action: TransactionAction;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type ListToolsParams = {
  status?: ListToolsStatus;
  category?: string;
};

export type ListToolsStatus =
  (typeof ListToolsStatus)[keyof typeof ListToolsStatus];

export const ListToolsStatus = {
  available: "available",
  issued: "issued",
  missing: "missing",
} as const;

export type ListTransactionsParams = {
  limit?: number;
};