# lib/api-zod/src/generated/types/issueInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface IssueInput {
  toolId: string;
  userId: string;
}