# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanSummary } from "./scanSummary";
import type { Tool } from "./tool";

export interface ScanResult {
  correct: Tool[];
  missing: Tool[];
  extra: string[];
  summary: ScanSummary;
}