# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalTools: number;
  availableTools: number;
  issuedTools: number;
  missingTools: number;
  recentTransactions: number;
}