# lib/api-zod/src/generated/types/toolInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolInputStatus } from "./toolInputStatus";

export interface ToolInput {
  /** @minLength 1 */
  toolId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  status?: ToolInputStatus;
}