# lib/api-zod/src/generated/types/toolUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolUpdateStatus } from "./toolUpdateStatus";

export interface ToolUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  status?: ToolUpdateStatus;
}